refactor(simulation): extract compiled loop and build paths
